refactor: split backend monolith into modules
