+++ backend/main.py
@@ -1,492 +1,77 @@
-import asyncio
-import hashlib
-import json
-import logging
-import os
-import time
-import uuid
-from collections import defaultdict
-from contextlib import asynccontextmanager
-from datetime import UTC, datetime, timedelta
-from pathlib import Path
-from typing import Annotated
+"""Application entry point: wires middleware, lifespan and routers together.
 
-import jwt
-from dotenv import load_dotenv
-from eth_account import Account
-from eth_account.messages import encode_defunct
-from eth_utils.address import to_checksum_address
-from fastapi import (
-    Depends,
-    FastAPI,
-    HTTPException,
-    Request,
-    WebSocket,
-    WebSocketDisconnect,
-)
+The backend is split into focused modules:
+
+* ``config``        — env parsing and static constants
+* ``schemas``       — Pydantic request bodies
+* ``dependencies``  — DB session + identity/admin auth dependencies
+* ``security``      — nonce hashing, JWT decoding, chat-signature verification
+* ``serializers``   — JSON payload builders shared by REST + WebSockets
+* ``services``      — game seeding, room teardown, background cleanup
+* ``ws.managers``   — WebSocket connection managers
+* ``routers.*``     — the HTTP/WebSocket endpoints, grouped by domain
+
+This module only assembles them.
+"""
+
+import asyncio
+import logging
+from contextlib import asynccontextmanager
+
+from fastapi import FastAPI, Request
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.security import OAuth2PasswordBearer
-from pydantic import BaseModel, HttpUrl
-from pydantic import Field as PydField
-from slowapi import Limiter
 from slowapi.errors import RateLimitExceeded
 from slowapi.extension import _rate_limit_exceeded_handler
-from slowapi.util import get_remote_address
-from sqlalchemy import func
-from sqlalchemy.exc import IntegrityError
-from sqlmodel import Session, select
 
-from database import engine, get_session
-from models import (
-    DEFAULT_LOBBY_LOCATION,
-    Game,
-    Message,
-    Nonce,
-    Room,
-    RoomEvent,
-    RoomEventRsvp,
-    RoomMember,
-    RsvpStatus,
-    User,
+# ``DEFAULT_RATE_LIMIT`` is used below by ``read_root``; the remaining names are
+# backward-compatible re-exports. The test-suite reaches into ``main`` as a
+# white-box facade (``main.X``), including in-place mutation of
+# ``ADMIN_ADDRESSES`` — re-exporting the same objects keeps those references
+# resolving after the split. See ``__all__`` for the full re-exported surface.
+from config import (  # noqa: F401
+    ADMIN_ADDRESSES,
+    DEFAULT_RATE_LIMIT,
+    SYSTEM_SENDER_ADDRESS,
+    _parse_admin_addresses,
+    is_admin_address,
 )
-from usernames import validate_username
+from database import get_session
+from rate_limit import limiter
+from routers import auth, chat, events, games, rooms, users
+from security import hash_nonce  # noqa: F401
+from serializers import (  # noqa: F401
+    _iso,
+    _members_payload,
+    _msg_dict,
+    get_rooms_payload,
+)
+from services import cleanup_expired_rooms_task, seed_default_games
+from ws.managers import (  # noqa: F401
+    ConnectionManager,
+    RoomChatManager,
+    chat_manager,
+    manager,
+)
+
+__all__ = [
+    "ADMIN_ADDRESSES",
+    "ConnectionManager",
+    "RoomChatManager",
+    "SYSTEM_SENDER_ADDRESS",
+    "_iso",
+    "_members_payload",
+    "_msg_dict",
+    "_parse_admin_addresses",
+    "app",
+    "chat_manager",
+    "get_rooms_payload",
+    "get_session",
+    "hash_nonce",
+    "is_admin_address",
+    "manager",
+]
 
 logger = logging.getLogger(__name__)
-
-# Load .env from project root if it exists
-env_path = Path(__file__).parent.parent / ".env"
-if env_path.exists():
-    load_dotenv(dotenv_path=env_path)
-
-# --- Configuration ---
-DATABASE_URL = os.getenv("DATABASE_URL")
-if not DATABASE_URL:
-    raise RuntimeError("DATABASE_URL environment variable is not set")
-
-JWT_SECRET = os.getenv("JWT_SECRET")
-if not JWT_SECRET:
-    raise RuntimeError("JWT_SECRET environment variable is not set")
-
-JWT_ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
-NONCE_EXPIRATION_MINUTES = int(os.getenv("NONCE_EXPIRATION_MINUTES", "5"))
-JWT_EXPIRATION_MINUTES = int(os.getenv("JWT_EXPIRATION_MINUTES", "60"))
-ROOM_CLEANUP_INTERVAL_SECONDS = int(os.getenv("ROOM_CLEANUP_INTERVAL_SECONDS", "60"))
-
-DEFAULT_RATE_LIMIT = os.getenv("DEFAULT_RATE_LIMIT", "10/minute")
-
-WS_LIMITS = defaultdict(list)
-
-MAX_MSGS = int(os.getenv("WS_RATE_LIMIT_MAX_MESSAGES", "10"))
-WINDOW = int(os.getenv("WS_RATE_LIMIT_TIME_WINDOW_SECONDS", "10"))
-
-
-def _parse_admin_addresses(raw: str | None) -> set[str]:
-    """Parse the comma-separated `ADMIN_ADDRESSES` env value into a set.
-
-    Addresses are lower-cased so membership checks are case-insensitive and
-    independent of EIP-55 checksum casing. An empty/unset value yields an
-    empty set — i.e. the system simply has no admins.
-    """
-    if not raw:
-        return set()
-    return {part.strip().lower() for part in raw.split(",") if part.strip()}
-
-
-# Whitelisted admin identity addresses. Authority lives entirely in the
-# environment — there is no `is_admin` column, so the only way to grant power
-# is to edit `.env`. Kept mutable at module scope so tests can adjust it.
-ADMIN_ADDRESSES: set[str] = _parse_admin_addresses(os.getenv("ADMIN_ADDRESSES"))
-
-
-def is_admin_address(address: str) -> bool:
-    """Return whether `address` is whitelisted as an administrator."""
-    return address.lower() in ADMIN_ADDRESSES
-
-
-# --- Schemas ---
-class VerifyRequest(BaseModel):
-    address: str
-    nonce: str
-    signature: str
-
-
-class CreateRoomRequest(BaseModel):
-    name: str
-    game: str = PydField(min_length=1, max_length=100)
-    lobby_location: str = PydField(min_length=1, max_length=50)
-    players_max: int
-    description: str | None = PydField(default=None, max_length=500)
-    communicator_link: HttpUrl | None = None
-    requirements: str | None = PydField(default=None, max_length=1000)
-
-
-class ScheduleEventRequest(BaseModel):
-    """Body for `PUT /rooms/{name}/event`. Both timestamps must be in the future."""
-
-    starts_at: datetime
-    ends_at: datetime
-
-
-class SetRsvpRequest(BaseModel):
-    """Body for `PUT /rooms/{name}/event/rsvp`."""
-
-    status: RsvpStatus
-
-
-class CreateGameRequest(BaseModel):
-    """Body for `POST /games` — admin adds a new game category."""
-
-    name: str = PydField(min_length=1, max_length=100)
-
-
-LOBBY_LOCATIONS: list[dict[str, str | float]] = [
-    {"code": "na-east", "label": "North America East", "lat": 39.0, "lon": -77.0},
-    {"code": "na-west", "label": "North America West", "lat": 37.8, "lon": -122.4},
-    {"code": "eu-west", "label": "Europe West", "lat": 51.5, "lon": -0.1},
-    {"code": "eu-central", "label": "Europe Central", "lat": 50.1, "lon": 8.7},
-    {"code": "eu-north", "label": "Europe North", "lat": 59.3, "lon": 18.1},
-    {"code": "sa-east", "label": "South America East", "lat": -23.6, "lon": -46.6},
-    {"code": "asia-east", "label": "Asia East", "lat": 35.7, "lon": 139.7},
-    {"code": "asia-south", "label": "Asia South", "lat": 1.3, "lon": 103.8},
-    {"code": "oceania", "label": "Oceania", "lat": -33.9, "lon": 151.2},
-    {"code": "africa-south", "label": "Africa South", "lat": -26.2, "lon": 28.0},
-]
-LOBBY_LOCATION_CODES = {str(location["code"]) for location in LOBBY_LOCATIONS}
-
-
-# --- WebSocket connection manager ---
-class ConnectionManager:
-    def __init__(self):
-        self.active_connections: list[WebSocket] = []
-
-    async def connect(self, websocket: WebSocket):
-        await websocket.accept()
-        self.active_connections.append(websocket)
-
-    def disconnect(self, websocket: WebSocket):
-        if websocket in self.active_connections:
-            self.active_connections.remove(websocket)
-
-    async def broadcast(self, data: str):
-        for connection in list(self.active_connections):
-            try:
-                await connection.send_text(data)
-            except Exception:
-                # A stale global-list listener must not turn an already
-                # committed room mutation into an HTTP 500 response.
-                self.disconnect(connection)
-
-
-manager = ConnectionManager()
-
-
-class RoomChatManager:
-    def __init__(self):
-        self.rooms: dict[str, dict[WebSocket, str]] = {}
-
-    async def connect(self, room: str, websocket: WebSocket, address: str):
-        await websocket.accept()
-        self.rooms.setdefault(room, {})[websocket] = address
-
-    def disconnect(self, room: str, websocket: WebSocket):
-        if room in self.rooms and websocket in self.rooms[room]:
-            del self.rooms[room][websocket]
-            if not self.rooms[room]:
-                del self.rooms[room]
-
-    async def broadcast(self, room: str, payload: str):
-        for connection in list(self.rooms.get(room, {})):
-            try:
-                await connection.send_text(payload)
-            except Exception:
-                # Browsers can disappear without completing a close handshake.
-                # Drop that socket and continue updating the remaining room.
-                self.disconnect(room, connection)
-
-    async def disconnect_user(self, room: str, address: str, code: int = 4409):
-        """Close every room-chat connection authenticated as ``address``."""
-        connections = self.rooms.get(room, {})
-        targets = [
-            websocket
-            for websocket, connected_address in connections.items()
-            if connected_address.lower() == address.lower()
-        ]
-        for websocket in targets:
-            try:
-                await websocket.close(code=code)
-            except RuntimeError:
-                # The browser may have closed between the broadcast and this
-                # targeted cleanup. Its registry entry still needs removing.
-                pass
-            finally:
-                self.disconnect(room, websocket)
-
-
-chat_manager = RoomChatManager()
-
-SYSTEM_SENDER_ADDRESS = "__system__"
-
-
-# --- Helpers ---
-def hash_nonce(nonce_value: str) -> str:
-    return hashlib.sha256(nonce_value.encode()).hexdigest()
-
-
-def get_rooms_payload(session: Session) -> str:
-    now = datetime.now(UTC)
-
-    # Optional cleanup of expired rooms before returning the payload.
-    # NOTE: The DB layer also has ON DELETE CASCADE on RoomEvent / RoomEventRsvp
-    # / Message, but we delete explicitly so the same code path works on
-    # SQLite (test environment) where FK cascades depend on PRAGMA settings.
-    expired_rooms = session.exec(select(Room).where(Room.expires_at <= now)).all()
-    for er in expired_rooms:
-        old_messages = session.exec(
-            select(Message).where(Message.room_id == er.id)
-        ).all()
-        for m in old_messages:
-            session.delete(m)
-        old_event = session.exec(
-            select(RoomEvent).where(RoomEvent.room_id == er.id)
-        ).first()
-        if old_event is not None:
-            old_rsvps = session.exec(
-                select(RoomEventRsvp).where(RoomEventRsvp.event_id == old_event.id)
-            ).all()
-            for r in old_rsvps:
-                session.delete(r)
-            session.delete(old_event)
-        session.delete(er)
-    if expired_rooms:
-        session.commit()
-
-    return json.dumps(
-        [
-            {
-                "name": r.name,
-                "game": r.game,
-                "lobby_location": r.lobby_location,
-                "players_active": len(r.members),
-                "players_max": r.players_max,
-                "member_addresses": [m.identity_address for m in r.members],
-                "description": r.description,
-                "communicator_link": r.communicator_link,
-                "requirements": r.requirements,
-                "expires_at": _iso(r.expires_at),
-            }
-            for r in session.exec(select(Room)).all()
-        ]
-    )
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/verify", auto_error=False)
-
-
-def _decode_jwt(token: str) -> str:
-    """Decode a JWT and return its `sub` claim. Raises on any failure."""
-    payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
-    address = payload.get("sub")
-    if not isinstance(address, str):
-        raise jwt.InvalidTokenError("Missing sub claim")
-    return address
-
-
-async def get_current_user_address(
-    token: Annotated[str | None, Depends(oauth2_scheme)],
-) -> str:
-    if not token:
-        raise HTTPException(status_code=401, detail="Not authenticated")
-    try:
-        return _decode_jwt(token)
-    except jwt.ExpiredSignatureError:
-        raise HTTPException(status_code=401, detail="Token expired") from None
-    except jwt.InvalidTokenError, KeyError:
-        raise HTTPException(status_code=401, detail="Invalid token") from None
-
-
-async def get_admin_address(
-    address: Annotated[str, Depends(get_current_user_address)],
-) -> str:
-    """Authorize the caller as an administrator.
-
-    Reuses the normal JWT auth to resolve the identity address, then checks it
-    against the `ADMIN_ADDRESSES` whitelist. The whitelist is authoritative —
-    a forged `is_admin` JWT claim cannot grant access here.
-    """
-    if not is_admin_address(address):
-        raise HTTPException(status_code=403, detail="Admin privileges required")
-    return address
-
-
-def _iso(dt: datetime) -> str:
-    """Serialize a datetime to ISO 8601, appending `Z` for naive UTC values."""
-    return dt.isoformat() + "Z" if dt.tzinfo is None else dt.isoformat()
-
-
-def _members_payload(room: Room) -> list[dict[str, str | bool]]:
-    """Compact roster for chat WS `roster_update` frames."""
-    return [
-        {
-            "address": m.identity_address,
-            "username": m.username,
-            "is_admin": is_admin_address(m.identity_address),
-        }
-        for m in room.members
-    ]
-
-
-def _serialize_event_state(session: Session, room: Room) -> dict | None:
-    """Build the JSON payload for a room's scheduled event.
-
-    Returns `None` when the room has no event yet. Used by both the REST
-    endpoints and the chat WebSocket so that `event_update` frames stay
-    in sync with `GET /rooms/{name}/event`.
-    """
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is None:
-        return None
-
-    rsvp_rows = session.exec(
-        select(RoomEventRsvp, User)
-        .join(User, User.id == RoomEventRsvp.user_id)
-        .where(RoomEventRsvp.event_id == event.id)
-    ).all()
-
-    return {
-        "starts_at": _iso(event.starts_at),
-        "ends_at": _iso(event.ends_at),
-        "created_by": event.created_by,
-        "created_at": _iso(event.created_at),
-        "updated_at": _iso(event.updated_at),
-        "rsvps": [
-            {
-                "address": user.identity_address,
-                "username": user.username,
-                "status": rsvp.status.value,
-                "updated_at": _iso(rsvp.updated_at),
-            }
-            for rsvp, user in rsvp_rows
-        ],
-    }
-
-
-def _ensure_game(session: Session, name: str) -> Game:
-    game = session.exec(select(Game).where(Game.name == name)).first()
-    if game is not None:
-        return game
-
-    next_order = 1 + max(
-        (g.sort_order for g in session.exec(select(Game)).all()), default=0
-    )
-    game = Game(name=name, sort_order=next_order)
-    session.add(game)
-    try:
-        session.commit()
-    except IntegrityError:
-        session.rollback()
-        existing = session.exec(select(Game).where(Game.name == name)).first()
-        if existing is not None:
-            return existing
-        raise HTTPException(status_code=409, detail="Game already exists") from None
-    session.refresh(game)
-    return game
-
-
-def _purge_room(session: Session, room: Room) -> None:
-    messages = session.exec(select(Message).where(Message.room_id == room.id)).all()
-    for m in messages:
-        session.delete(m)
-
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is not None:
-        rsvps = session.exec(
-            select(RoomEventRsvp).where(RoomEventRsvp.event_id == event.id)
-        ).all()
-        for r in rsvps:
-            session.delete(r)
-        session.delete(event)
-
-    room.members.clear()
-    session.delete(room)
-
-
-DEFAULT_GAMES: list[str] = [
-    "Quake III Arena",
-    "Diablo II",
-    "StarCraft",
-    "Half-Life",
-    "Unreal Tournament",
-]
-
-
-def seed_default_games() -> None:
-    from database import engine
-
-    with Session(engine) as session:
-        existing = {g.name for g in session.exec(select(Game)).all()}
-        next_order = 1 + max(
-            (g.sort_order for g in session.exec(select(Game)).all()), default=0
-        )
-        added = False
-        for name in DEFAULT_GAMES:
-            if name in existing:
-                continue
-            session.add(Game(name=name, sort_order=next_order))
-            next_order += 1
-            added = True
-        if added:
-            session.commit()
-
-
-async def cleanup_expired_rooms_task() -> None:
-    """Periodically clean up expired rooms from the database."""
-    while True:
-        try:
-            await asyncio.sleep(ROOM_CLEANUP_INTERVAL_SECONDS)
-
-            with Session(engine) as session:
-                now = datetime.now(UTC)
-                expired_rooms = session.exec(
-                    select(Room).where(Room.expires_at <= now)
-                ).all()
-
-                total_messages_deleted = 0
-                for er in expired_rooms:
-                    old_messages = session.exec(
-                        select(Message).where(Message.room_id == er.id)
-                    ).all()
-                    total_messages_deleted += len(old_messages)
-                    for m in old_messages:
-                        session.delete(m)
-                    old_event = session.exec(
-                        select(RoomEvent).where(RoomEvent.room_id == er.id)
-                    ).first()
-                    if old_event is not None:
-                        old_rsvps = session.exec(
-                            select(RoomEventRsvp).where(
-                                RoomEventRsvp.event_id == old_event.id
-                            )
-                        ).all()
-                        for r in old_rsvps:
-                            session.delete(r)
-                        session.delete(old_event)
-                    session.delete(er)
-
-                if expired_rooms:
-                    session.commit()
-                    rooms_count = len(expired_rooms)
-                    logger.info(
-                        f"""Cleanup task deleted {rooms_count} room(s)
-                        and {total_messages_deleted} message(s)"""
-                    )
-
-                    # Broadcast updated rooms payload to all WebSocket clients
-                    rooms_payload = get_rooms_payload(session)
-                    await manager.broadcast(rooms_payload)
-        except Exception as exc:
-            logger.exception(f"Error in cleanup_expired_rooms_task: {exc}")
 
 
 @asynccontextmanager
@@ -504,8 +89,6 @@ async def lifespan(_app: FastAPI):
 
 
 app = FastAPI(lifespan=lifespan)
-
-limiter = Limiter(key_func=get_remote_address)
 
 app.state.limiter = limiter
 app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)
@@ -526,8 +109,6 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# --- Endpoints ---
-
 
 @app.get("/")
 @limiter.limit(DEFAULT_RATE_LIMIT)
@@ -535,1008 +116,9 @@ def read_root(request: Request):  # noqa: ARG001
     return {"status": "ok", "service": "playlink-auth"}
 
 
-@app.post("/auth/request-nonce")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def request_nonce(request: Request, address: str, session: SessionDep):  # noqa: ARG001
-    """
-    Request a one-time nonce for an identity address.
-    """
-    try:
-        checksum_address = to_checksum_address(address)
-    except ValueError:
-        raise HTTPException(
-            status_code=400, detail="Invalid identity address format"
-        ) from None
-
-    # Upsert User
-    user = session.exec(
-        select(User).where(User.identity_address == checksum_address)
-    ).first()
-    if not user:
-        user = User(identity_address=checksum_address)
-        session.add(user)
-        session.commit()
-        session.refresh(user)
-
-    # Generate & Hash Nonce
-    nonce_value = str(uuid.uuid4())
-    hashed_value = hash_nonce(nonce_value)
-
-    # Invalidate previous unused nonces for this identity
-    previous_nonces = session.exec(
-        select(Nonce).where(
-            Nonce.identity_address == checksum_address,
-            Nonce.used == False,  # noqa: E712
-        )
-    ).all()
-    for old_nonce in previous_nonces:
-        old_nonce.used = True
-        session.add(old_nonce)
-
-    # Store new hashed nonce
-    expires_at = datetime.now(UTC) + timedelta(minutes=NONCE_EXPIRATION_MINUTES)
-    db_nonce = Nonce(
-        identity_address=checksum_address,
-        value=hashed_value,
-        expires_at=expires_at,
-        user_id=user.id,
-    )
-    session.add(db_nonce)
-    session.commit()
-
-    return {"nonce": nonce_value}
-
-
-@app.post("/auth/verify")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def verify_signature(request: Request, body: VerifyRequest, session: SessionDep):  # noqa: ARG001
-    """
-    Verify signature against a nonce and issue a JWT.
-    """
-    address = body.address
-    nonce = body.nonce
-    signature = body.signature
-
-    try:
-        checksum_address = to_checksum_address(address)
-    except ValueError:
-        raise HTTPException(
-            status_code=400, detail="Invalid identity address format"
-        ) from None
-
-    hashed_provided_nonce = hash_nonce(nonce)
-
-    # Fetch matching unused & unexpired nonce
-    db_nonce = session.exec(
-        select(Nonce).where(
-            Nonce.identity_address == checksum_address,
-            Nonce.value == hashed_provided_nonce,
-            Nonce.used == False,  # noqa: E712
-            Nonce.expires_at > datetime.now(UTC),
-        )
-    ).first()
-
-    if not db_nonce:
-        raise HTTPException(status_code=401, detail="Invalid or expired challenge")
-
-    # Recover address from signature
-    message_text = f"Sign in to Playlink\nNonce: {nonce}"
-    message = encode_defunct(text=message_text)
-
-    try:
-        recovered_address = Account.recover_message(message, signature=signature)
-    except Exception:
-        raise HTTPException(
-            status_code=401, detail="Invalid signature format"
-        ) from None
-
-    if recovered_address.lower() != checksum_address.lower():
-        raise HTTPException(status_code=401, detail="Identity verification failed")
-
-    # Success: Mark nonce as used & update user
-    db_nonce.used = True
-    session.add(db_nonce)
-
-    user = session.get(User, db_nonce.user_id)
-    if user:
-        user.last_login = datetime.now(UTC)
-        session.add(user)
-
-    session.commit()
-
-    # Generate JWT
-    is_admin = is_admin_address(checksum_address)
-    token_data = {
-        "sub": checksum_address,
-        "username": user.username if user else "Unknown",
-        "is_admin": is_admin,
-        "iat": datetime.now(UTC),
-        "exp": datetime.now(UTC) + timedelta(minutes=JWT_EXPIRATION_MINUTES),
-        "iss": "playlink-auth",
-    }
-    token = jwt.encode(token_data, JWT_SECRET, algorithm=JWT_ALGORITHM)
-
-    return {
-        "token": token,
-        "username": user.username if user else "Unknown",
-        "is_admin": is_admin,
-    }
-
-
-def _user_payload(user: User) -> dict:
-    """Public-facing serialization of a user (shared by GET/PATCH /users/me)."""
-    return {
-        "id": user.id,
-        "identity_address": user.identity_address,
-        "username": user.username,
-        "created_at": _iso(user.created_at),
-        "last_login": _iso(user.last_login) if user.last_login else None,
-        "is_admin": is_admin_address(user.identity_address),
-    }
-
-
-class UpdateUserRequest(BaseModel):
-    username: str
-
-
-@app.get("/users/me")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def get_me(
-    request: Request,  # noqa: ARG001
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    return _user_payload(user)
-
-
-@app.patch("/users/me")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def update_me(
-    request: Request,  # noqa: ARG001
-    payload: UpdateUserRequest,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    username = payload.username.strip()
-
-    error = validate_username(username)
-    if error == "invalid_format":
-        raise HTTPException(
-            status_code=400,
-            detail="Username must be 3-20 characters: letters, numbers, _ or -.",
-        )
-    if error == "profane":
-        raise HTTPException(
-            status_code=400, detail="Username contains inappropriate language."
-        )
-
-    if username == user.username:
-        return _user_payload(user)
-
-    existing = session.exec(select(User).where(User.username == username)).first()
-    if existing is not None:
-        raise HTTPException(status_code=409, detail="Username already taken.")
-
-    user.username = username
-    session.add(user)
-    try:
-        session.commit()
-    except IntegrityError:
-        session.rollback()
-        raise HTTPException(status_code=409, detail="Username already taken.") from None
-    session.refresh(user)
-    return _user_payload(user)
-
-
-@app.get("/rooms")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def list_rooms(request: Request, session: SessionDep):  # noqa: ARG001
-    return session.exec(select(Room)).all()
-
-
-@app.get("/lobby-locations")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def list_lobby_locations(request: Request):  # noqa: ARG001
-    return {"default": DEFAULT_LOBBY_LOCATION, "locations": LOBBY_LOCATIONS}
-
-
-@app.get("/rooms/{room_name}")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def get_room(request: Request, room_name: str, session: SessionDep):  # noqa: ARG001
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-    return {
-        "name": room.name,
-        "game": room.game,
-        "lobby_location": room.lobby_location,
-        "players_max": room.players_max,
-        "players_active": len(room.members),
-        "member_addresses": [m.identity_address for m in room.members],
-        "members": _members_payload(room),
-        "description": room.description,
-        "communicator_link": room.communicator_link,
-        "requirements": room.requirements,
-        "created_by": room.created_by,
-        "expires_at": _iso(room.expires_at),
-        "event": _serialize_event_state(session, room),
-    }
-
-
-@app.get("/games")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def list_games(request: Request, session: SessionDep):  # noqa: ARG001
-    games = session.exec(select(Game).order_by(Game.sort_order)).all()
-    return [game.name for game in games]
-
-
-@app.delete(
-    "/rooms/{room_name}",
-    status_code=200,
-    dependencies=[Depends(get_admin_address)],
-)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def delete_room(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    session: SessionDep,
-):
-    """Admin-only: close a room.
-
-    Cascades deletion of the room's chat messages, scheduled event and RSVPs,
-    notifies anyone connected to the room's chat via a `room_closed` frame so
-    their client can redirect them out, and broadcasts the refreshed active
-    rooms list to all global listeners.
-    """
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    _purge_room(session, room)
-    session.commit()
-
-    await chat_manager.broadcast(
-        room_name, json.dumps({"type": "room_closed", "room": room_name})
-    )
-    await manager.broadcast(get_rooms_payload(session))
-    return {"status": "closed", "room": room_name}
-
-
-@app.post("/games", status_code=201, dependencies=[Depends(get_admin_address)])
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def create_game(
-    request: Request,  # noqa: ARG001
-    body: CreateGameRequest,
-    session: SessionDep,
-):
-    """Admin-only: add a new game category at the end of the sort order."""
-    name = body.name.strip()
-    if not name:
-        raise HTTPException(status_code=400, detail="Game name is required")
-
-    existing = session.exec(select(Game).where(Game.name == name)).first()
-    if existing:
-        raise HTTPException(status_code=409, detail="Game already exists")
-
-    next_order = 1 + max(
-        (g.sort_order for g in session.exec(select(Game)).all()), default=0
-    )
-    game = Game(name=name, sort_order=next_order)
-    session.add(game)
-    session.commit()
-    session.refresh(game)
-    return {"name": game.name, "sort_order": game.sort_order}
-
-
-@app.delete("/games/{name}", status_code=200, dependencies=[Depends(get_admin_address)])
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def delete_game(
-    request: Request,  # noqa: ARG001
-    name: str,
-    session: SessionDep,
-    force: bool = False,
-):
-    """Admin-only: delete a game category.
-
-    If rooms are still playing this game the request is refused with `409`
-    unless `force=true`, in which case those rooms are closed first (cascading
-    like `DELETE /rooms/{name}`) and their chat clients are redirected out.
-    """
-    game = session.exec(select(Game).where(Game.name == name)).first()
-    if not game:
-        raise HTTPException(status_code=404, detail="Game not found")
-
-    rooms = session.exec(select(Room).where(Room.game == name)).all()
-    if rooms and not force:
-        raise HTTPException(
-            status_code=409,
-            detail=f"There are {len(rooms)} active rooms currently playing this game.",
-        )
-
-    closed_rooms = [room.name for room in rooms]
-    for room in rooms:
-        _purge_room(session, room)
-    session.delete(game)
-    session.commit()
-
-    for closed_name in closed_rooms:
-        await chat_manager.broadcast(
-            closed_name, json.dumps({"type": "room_closed", "room": closed_name})
-        )
-    if closed_rooms:
-        await manager.broadcast(get_rooms_payload(session))
-    return {"status": "deleted", "game": name, "rooms_closed": closed_rooms}
-
-
-@app.post("/rooms", status_code=201)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def create_room(
-    request: Request,  # noqa: ARG001
-    body: CreateRoomRequest,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    now = datetime.now(UTC)
-
-    # Clean up expired rooms for accurate counts
-    expired_rooms = session.exec(select(Room).where(Room.expires_at <= now)).all()
-    for er in expired_rooms:
-        old_messages = session.exec(
-            select(Message).where(Message.room_id == er.id)
-        ).all()
-        for m in old_messages:
-            session.delete(m)
-        session.delete(er)
-    session.commit()
-
-    existing = session.exec(select(Room).where(Room.name == body.name)).first()
-    if existing:
-        raise HTTPException(status_code=409, detail="Room name already taken")
-
-    if not is_admin_address(address):
-        user_rooms_count = len(
-            session.exec(select(Room).where(Room.created_by == address)).all()
-        )
-        if user_rooms_count >= 3:
-            raise HTTPException(
-                status_code=400, detail="You can create a maximum of 3 rooms."
-            )
-
-    game_name = body.game.strip()
-    if not game_name:
-        raise HTTPException(status_code=400, detail="Game name is required")
-
-    lobby_location = body.lobby_location.strip()
-    if lobby_location not in LOBBY_LOCATION_CODES:
-        raise HTTPException(status_code=400, detail="Unsupported lobby location")
-
-    _ = _ensure_game(session, game_name)
-
-    room = Room(
-        name=body.name,
-        game=game_name,
-        lobby_location=lobby_location,
-        players_max=body.players_max,
-        description=body.description,
-        communicator_link=(
-            str(body.communicator_link) if body.communicator_link else None
-        ),
-        requirements=body.requirements,
-        created_by=address,
-    )
-
-    # Auto-join creator
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if user:
-        room.members.append(user)
-
-    session.add(room)
-    session.commit()
-    session.refresh(room)
-
-    await manager.broadcast(get_rooms_payload(session))
-    return {"status": "created", "room": room.name}
-
-
-@app.post("/rooms/{room_name}/join", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def join_room(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    if user in room.members:
-        raise HTTPException(status_code=400, detail="You are already in this room")
-
-    if len(room.members) >= room.players_max:
-        raise HTTPException(status_code=400, detail="Room is full")
-
-    room.members.append(user)
-    session.add(room)
-    session.commit()
-
-    await manager.broadcast(get_rooms_payload(session))
-    await chat_manager.broadcast(
-        room_name,
-        json.dumps({"type": "roster_update", "members": _members_payload(room)}),
-    )
-    return {"status": "joined", "room": room.name}
-
-
-@app.post("/rooms/{room_name}/leave", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def leave_room(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    if user not in room.members:
-        raise HTTPException(status_code=400, detail="You are not in this room")
-
-    # Drop the leaving user's RSVP (if any) so an event's roster reflects
-    # the current member set. Idempotent — runs even when no event exists.
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    rsvp_to_drop: RoomEventRsvp | None = None
-    if event is not None:
-        rsvp_to_drop = session.exec(
-            select(RoomEventRsvp).where(
-                RoomEventRsvp.event_id == event.id,
-                RoomEventRsvp.user_id == user.id,
-            )
-        ).first()
-        if rsvp_to_drop is not None:
-            session.delete(rsvp_to_drop)
-
-    room.members.remove(user)
-    session.add(room)
-    session.commit()
-
-    await manager.broadcast(get_rooms_payload(session))
-    await chat_manager.broadcast(
-        room_name,
-        json.dumps({"type": "roster_update", "members": _members_payload(room)}),
-    )
-    if rsvp_to_drop is not None:
-        await chat_manager.broadcast(
-            room_name,
-            json.dumps(
-                {"type": "event_update", "event": _serialize_event_state(session, room)}
-            ),
-        )
-    return {"status": "left", "room": room.name}
-
-
-@app.post("/rooms/{room_name}/members/{member_address}/kick", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def kick_room_member(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    member_address: str,
-    session: SessionDep,
-    admin_address: Annotated[str, Depends(get_admin_address)],
-):
-    """Remove a non-admin member and disconnect their active chat sessions."""
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    target = session.exec(
-        select(User).where(func.lower(User.identity_address) == member_address.lower())
-    ).first()
-    if not target:
-        raise HTTPException(status_code=404, detail="User not found")
-    if is_admin_address(target.identity_address):
-        raise HTTPException(status_code=403, detail="Administrators cannot be kicked")
-    if target not in room.members:
-        raise HTTPException(status_code=409, detail="User is not a room member")
-
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is not None:
-        rsvp = session.exec(
-            select(RoomEventRsvp).where(
-                RoomEventRsvp.event_id == event.id,
-                RoomEventRsvp.user_id == target.id,
-            )
-        ).first()
-        if rsvp is not None:
-            session.delete(rsvp)
-
-    ownership_transferred = room.created_by.lower() == target.identity_address.lower()
-    if ownership_transferred:
-        room.created_by = admin_address
-
-    room.members.remove(target)
-    system_message = Message(
-        room_id=room.id,
-        sender_address=SYSTEM_SENDER_ADDRESS,
-        content=(f"{target.username} was removed from the room by an administrator."),
-    )
-    session.add(room)
-    session.add(system_message)
-    session.commit()
-    session.refresh(system_message)
-
-    await chat_manager.broadcast(
-        room_name,
-        json.dumps(
-            {
-                "type": "message",
-                "message": _msg_dict(system_message, "System"),
-            }
-        ),
-    )
-    await chat_manager.broadcast(
-        room_name,
-        json.dumps({"type": "roster_update", "members": _members_payload(room)}),
-    )
-    if event is not None:
-        await chat_manager.broadcast(
-            room_name,
-            json.dumps(
-                {"type": "event_update", "event": _serialize_event_state(session, room)}
-            ),
-        )
-    await chat_manager.broadcast(
-        room_name,
-        json.dumps(
-            {
-                "type": "member_kicked",
-                "member_address": target.identity_address,
-                "member_username": target.username,
-                "created_by": room.created_by,
-                "ownership_transferred": ownership_transferred,
-            }
-        ),
-    )
-    await manager.broadcast(get_rooms_payload(session))
-    await chat_manager.disconnect_user(room_name, target.identity_address)
-
-    return {
-        "status": "kicked",
-        "member_address": target.identity_address,
-        "member_username": target.username,
-        "created_by": room.created_by,
-        "ownership_transferred": ownership_transferred,
-    }
-
-
-@app.get("/rooms/{room_name}/event")
-@limiter.limit(DEFAULT_RATE_LIMIT)
-def get_room_event(request: Request, room_name: str, session: SessionDep):  # noqa: ARG001
-    """Return the room's scheduled event (if any), including the RSVP roster.
-
-    Mirrors `GET /rooms/{name}` in being public — anyone can browse the
-    schedule, only members are allowed to RSVP.
-    """
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-    event_state = _serialize_event_state(session, room)
-    if event_state is None:
-        raise HTTPException(status_code=404, detail="No event scheduled")
-    return event_state
-
-
-@app.put("/rooms/{room_name}/event", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def schedule_room_event(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    body: ScheduleEventRequest,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    """Create or replace the scheduled event for a room.
-
-    Only the room creator may schedule. `starts_at` must lie in the future
-    and `ends_at` must come after `starts_at`. The room's `expires_at` is
-    automatically extended past `ends_at` so the room stays alive until the
-    event finishes (plus a small grace period for stragglers).
-    """
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    if room.created_by.lower() != address.lower():
-        raise HTTPException(
-            status_code=403, detail="Only the room creator can schedule an event"
-        )
-
-    starts_at = body.starts_at
-    ends_at = body.ends_at
-    if starts_at.tzinfo is None:
-        starts_at = starts_at.replace(tzinfo=UTC)
-    else:
-        starts_at = starts_at.astimezone(UTC)
-    if ends_at.tzinfo is None:
-        ends_at = ends_at.replace(tzinfo=UTC)
-    else:
-        ends_at = ends_at.astimezone(UTC)
-
-    now = datetime.now(UTC)
-    if starts_at <= now:
-        raise HTTPException(status_code=400, detail="starts_at must be in the future")
-    if ends_at <= starts_at:
-        raise HTTPException(status_code=400, detail="ends_at must be after starts_at")
-
-    # Keep the room alive for the entire event plus a 30-minute grace period
-    # so members don't get prune'd out mid-session.
-    grace = timedelta(minutes=30)
-    expires_at = room.expires_at
-    if expires_at.tzinfo is None:
-        expires_at = expires_at.replace(tzinfo=UTC)
-    new_expiry = max(expires_at, ends_at + grace)
-    room_expiry_changed = new_expiry != expires_at
-    if room_expiry_changed:
-        room.expires_at = new_expiry
-        session.add(room)
-
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is None:
-        event = RoomEvent(
-            room_id=room.id,
-            starts_at=starts_at,
-            ends_at=ends_at,
-            created_by=address,
-        )
-        session.add(event)
-    else:
-        # If the time window shifted at all, every RSVP was a decision about
-        # the old time — drop them so members re-confirm against the new slot.
-        existing_starts = event.starts_at
-        existing_ends = event.ends_at
-        if existing_starts.tzinfo is None:
-            existing_starts = existing_starts.replace(tzinfo=UTC)
-        if existing_ends.tzinfo is None:
-            existing_ends = existing_ends.replace(tzinfo=UTC)
-        time_changed = existing_starts != starts_at or existing_ends != ends_at
-        if time_changed:
-            stale_rsvps = session.exec(
-                select(RoomEventRsvp).where(RoomEventRsvp.event_id == event.id)
-            ).all()
-            for rsvp in stale_rsvps:
-                session.delete(rsvp)
-        event.starts_at = starts_at
-        event.ends_at = ends_at
-        event.updated_at = datetime.now(UTC)
-        session.add(event)
-    session.commit()
-
-    payload = _serialize_event_state(session, room)
-    await chat_manager.broadcast(
-        room_name, json.dumps({"type": "event_update", "event": payload})
-    )
-    if room_expiry_changed:
-        await manager.broadcast(get_rooms_payload(session))
-    return payload
-
-
-@app.delete("/rooms/{room_name}/event", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def cancel_room_event(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    """Delete the scheduled event for a room and clear all RSVPs."""
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    if room.created_by.lower() != address.lower():
-        raise HTTPException(
-            status_code=403, detail="Only the room creator can cancel the event"
-        )
-
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is None:
-        raise HTTPException(status_code=404, detail="No event scheduled")
-
-    rsvps = session.exec(
-        select(RoomEventRsvp).where(RoomEventRsvp.event_id == event.id)
-    ).all()
-    for r in rsvps:
-        session.delete(r)
-    session.delete(event)
-    session.commit()
-
-    await chat_manager.broadcast(
-        room_name, json.dumps({"type": "event_update", "event": None})
-    )
-    return {"status": "cancelled", "room": room.name}
-
-
-@app.put("/rooms/{room_name}/event/rsvp", status_code=200)
-@limiter.limit(DEFAULT_RATE_LIMIT)
-async def set_room_event_rsvp(
-    request: Request,  # noqa: ARG001
-    room_name: str,
-    body: SetRsvpRequest,
-    session: SessionDep,
-    address: Annotated[str, Depends(get_current_user_address)],
-):
-    """Upsert the caller's RSVP for the room's scheduled event.
-
-    Only current members of the room may RSVP. The caller's status
-    overwrites any previous one (one RSVP per user per event).
-    """
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        raise HTTPException(status_code=404, detail="Room not found")
-
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    if user not in room.members:
-        raise HTTPException(
-            status_code=403, detail="Only room members can RSVP to the event"
-        )
-
-    event = session.exec(select(RoomEvent).where(RoomEvent.room_id == room.id)).first()
-    if event is None:
-        raise HTTPException(status_code=404, detail="No event scheduled")
-
-    rsvp = session.exec(
-        select(RoomEventRsvp).where(
-            RoomEventRsvp.event_id == event.id,
-            RoomEventRsvp.user_id == user.id,
-        )
-    ).first()
-    if rsvp is None:
-        rsvp = RoomEventRsvp(
-            event_id=event.id,
-            user_id=user.id,
-            status=body.status,
-        )
-        session.add(rsvp)
-    else:
-        rsvp.status = body.status
-        rsvp.updated_at = datetime.now(UTC)
-        session.add(rsvp)
-    session.commit()
-    session.refresh(rsvp)
-
-    rsvp_payload = {
-        "address": user.identity_address,
-        "username": user.username,
-        "status": rsvp.status.value,
-        "updated_at": _iso(rsvp.updated_at),
-    }
-    await chat_manager.broadcast(
-        room_name, json.dumps({"type": "rsvp_update", "rsvp": rsvp_payload})
-    )
-    return rsvp_payload
-
-
-@app.websocket("/ws/rooms")
-async def websocket_rooms(websocket: WebSocket, session: SessionDep):
-    await manager.connect(websocket)
-    try:
-        await websocket.send_text(get_rooms_payload(session))
-        while True:
-            await websocket.receive_text()
-    except WebSocketDisconnect:
-        manager.disconnect(websocket)
-
-
-# Issue #59: chat messages are signed with the sender's BIP39-derived key
-# (EIP-191 personal_sign) so authorship is verifiable end-to-end rather than
-# trusted on the JWT alone. Reject timestamps drifting more than this from the
-# server clock to blunt replay of captured signatures.
-CHAT_SIGNATURE_SKEW_SECONDS = 120
-
-
-def _chat_signing_message(room_name: str, content: str, sent_at: str) -> str:
-    """Canonical text the client signs and the server reconstructs.
-
-    Must stay byte-for-byte in sync with the frontend builder in
-    ``frontend/src/lib/signing.ts``.
-    """
-    return (
-        "PlayLink signed chat message\n"
-        f"room={room_name}\n"
-        f"sent_at={sent_at}\n"
-        f"content={content}"
-    )
-
-
-def _verify_chat_signature(
-    room_name: str, content: str, sent_at: str, signature: str, address: str
-) -> bool:
-    """True iff ``signature`` was produced by ``address`` over the canonical
-    payload and ``sent_at`` is within the allowed clock skew."""
-    try:
-        signed_at = datetime.fromisoformat(sent_at)
-    except ValueError, TypeError:
-        return False
-    if signed_at.tzinfo is None:
-        signed_at = signed_at.replace(tzinfo=UTC)
-    drift = abs((datetime.now(UTC) - signed_at).total_seconds())
-    if drift > CHAT_SIGNATURE_SKEW_SECONDS:
-        return False
-
-    message = encode_defunct(text=_chat_signing_message(room_name, content, sent_at))
-    try:
-        recovered = Account.recover_message(message, signature=signature)
-    except Exception:
-        return False
-    return recovered.lower() == address.lower()
-
-
-def _msg_dict(msg: Message, sender_username: str) -> dict:
-    payload = {
-        "id": msg.id,
-        "sender_address": msg.sender_address,
-        "sender_username": sender_username,
-        "content": msg.content,
-        "created_at": _iso(msg.created_at),
-        "signature": msg.signature,
-        "sent_at": msg.sent_at,
-        "verified": msg.signature is not None,
-    }
-    if msg.sender_address == SYSTEM_SENDER_ADDRESS:
-        payload["kind"] = "system"
-    return payload
-
-
-@app.websocket("/ws/rooms/{room_name}/chat")
-async def websocket_chat(
-    websocket: WebSocket,
-    room_name: str,
-    token: str,
-    session: SessionDep,
-):
-    # 1. Authenticate via JWT in query param (browsers can't set WS headers).
-    try:
-        address = _decode_jwt(token)
-    except Exception:
-        await websocket.close(code=4401)
-        return
-
-    # 2. Room must exist.
-    room = session.exec(select(Room).where(Room.name == room_name)).first()
-    if not room:
-        await websocket.close(code=4404)
-        return
-
-    # 3. Caller must be a member.
-    user = session.exec(select(User).where(User.identity_address == address)).first()
-    if not user or user not in room.members:
-        await websocket.close(code=4403)
-        return
-
-    await chat_manager.connect(room_name, websocket, address)
-    try:
-        # Send last 50 messages in chronological order.
-        recent = session.exec(
-            select(Message)
-            .where(Message.room_id == room.id)
-            .order_by(Message.created_at.desc())
-            .limit(50)
-        ).all()
-        username_cache: dict[str, str] = {}
-
-        def _username_for(addr: str) -> str:
-            if addr == SYSTEM_SENDER_ADDRESS:
-                return "System"
-            cached = username_cache.get(addr)
-            if cached is not None:
-                return cached
-            sender = session.exec(
-                select(User).where(User.identity_address == addr)
-            ).first()
-            name = sender.username if sender else addr
-            username_cache[addr] = name
-            return name
-
-        history_payload = json.dumps(
-            {
-                "type": "history",
-                "messages": [
-                    _msg_dict(m, _username_for(m.sender_address))
-                    for m in reversed(recent)
-                ],
-            }
-        )
-        await websocket.send_text(history_payload)
-
-        while True:
-            raw = await websocket.receive_text()
-
-            # Membership may have changed after the socket was accepted. Query
-            # the link table directly so a stale relationship cache cannot let
-            # a removed member continue writing to chat.
-            membership = session.exec(
-                select(RoomMember).where(
-                    RoomMember.room_id == room.id,
-                    RoomMember.user_id == user.id,
-                )
-            ).first()
-            if membership is None:
-                await websocket.close(code=4403)
-                return
-
-            now = time.time()
-
-            key = address
-
-            WS_LIMITS[key] = [t for t in WS_LIMITS[key] if now - t < WINDOW]
-
-            if len(WS_LIMITS[key]) >= MAX_MSGS:
-                await websocket.close(code=4429)
-                return
-
-            WS_LIMITS[key].append(now)
-
-            try:
-                data = json.loads(raw)
-                content = str(data.get("content", "")).strip()
-            except ValueError, TypeError:
-                continue
-            if not content or len(content) > 1000:
-                continue
-
-            # Issue #59: when a signature is supplied it must verify against the
-            # connected identity, otherwise the message is rejected outright
-            # (never stored or broadcast). A missing signature is a legacy /
-            # no-key client and is accepted as unverified.
-            raw_signature = data.get("signature")
-            signature: str | None = None
-            sent_at: str | None = None
-            if raw_signature:
-                signature = str(raw_signature)
-                sent_at = str(data.get("sent_at", ""))
-                if not _verify_chat_signature(
-                    room_name, content, sent_at, signature, address
-                ):
-                    await websocket.send_text(
-                        json.dumps({"type": "error", "detail": "signature_invalid"})
-                    )
-                    continue
-
-            msg = Message(
-                room_id=room.id,
-                sender_address=address,
-                content=content,
-                signature=signature,
-                sent_at=sent_at,
-            )
-            session.add(msg)
-            session.commit()
-            session.refresh(msg)
-
-            await chat_manager.broadcast(
-                room_name,
-                json.dumps(
-                    {"type": "message", "message": _msg_dict(msg, user.username)}
-                ),
-            )
-    except WebSocketDisconnect:
-        chat_manager.disconnect(room_name, websocket)
+app.include_router(auth.router)
+app.include_router(users.router)
+app.include_router(rooms.router)
+app.include_router(games.router)
+app.include_router(events.router)
+app.include_router(chat.router)
